Create pulse_processing_fast_sampling.py
This is the plugin producing records from only raw_records_v1730. It has the same structure of the pulse_processing.py plugin but we keep it separate in case we would need different records operation. At the current time I could test it on some raw data with the amstraxer.py and it worked.