Plugins review (part 2) (#119)
* Update pulse_processing.py

New pulse processing for XAMS data. For now the only difference is in a new baseline method "baseline_per_channel", which only flips the data coming from the pmt.

* Update peak_processing.py

Restore peak processing for XAMS data, leave commented the plugins that we are not using right now, as for example the Peaks Classification or the Peaks Positions.

* Create peak_basics.py

For now we keep it separately from radon_peak_basics because the area_fraction_top variable is computed differently.
Also, the n_top variable is hard coded but we will have it saved in the common configs soon.

* Update contexts.py

Separate the datastructure of xams and xamsl. Hopefully, for the best,

* Update contexts.py

Mistake in the name

* Update contexts.py

typo again

* Update __init__.py

forgot an import

* Update peak_basics.py

* Update pulse_processing.py

we will hard code this for now.

* Update peak_processing.py

* Update peak_processing.py